# Conflicts:
#	easyclass-repository/src/main/java/edu/vero/easyclass/repositories/CoursewareJpaDao.java
#	easyclass-repository/src/main/java/edu/vero/easyclass/repositories/TestRecordJpaDao.java
#	easyclass-service/src/main/java/edu/vero/easyclass/services/ClassTimeService.java
#	easyclass-service/src/main/java/edu/vero/easyclass/services/DepartmentService.java
#	easyclass-service/src/main/java/edu/vero/easyclass/services/HomeworkService.java
#	easyclass-service/src/main/java/edu/vero/easyclass/services/TeacherArrangementService.java
#	easyclass-service/src/main/java/edu/vero/easyclass/services/impl/ClassTimeServiceImpl.java
#	easyclass-service/src/main/java/edu/vero/easyclass/services/impl/DepartmentServiceImpl.java
#	easyclass-service/src/main/java/edu/vero/easyclass/services/impl/TeacherArrangementServiceImpl.java

+++ easyclass-repository/src/main/java/edu/vero/easyclass/repositories/UserJpaDao.java
@@ -14,6 +14,6 @@ import org.springframework.stereotype.Repository;
  */
 
 @Repository
-public interface UserJpaDao extends JpaRepository<User, Long> {
+public interface UserJpaDao extends JpaRepository<User, Integer> {
 
 }
